requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
jinja2==3.1.4
python-multipart==0.0.9

SQLAlchemy==2.0.32
pydantic==2.8.2

openai==1.35.10     # optional: if using OpenAI
groq==0.9.0         # optional: if using Groq

python-dotenv==1.0.1
